# src/main/resources/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Biblioteca",
    "description": "API REST para la gestión de libros de una biblioteca. Permite registrar, consultar, actualizar y eliminar libros, aplicando reglas de unicidad por ISBN y validaciones de entrada.",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Biblioteca",
      "email": "biblioteca@catolica.edu.sv"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de desarrollo local"
    }
  ],
  "tags": [
    {
      "name": "Libros",
      "description": "Operaciones sobre el catálogo de libros"
    }
  ],
  "paths": {
    "/api/v1/libros": {
      "get": {
        "tags": ["Libros"],
        "summary": "Listar libros",
        "description": "Devuelve la lista completa de libros. Permite filtrar opcionalmente por disponibilidad.",
        "parameters": [
          {
            "name": "disponible",
            "in": "query",
            "required": false,
            "description": "Si se envía true devuelve solo libros disponibles; false solo los no disponibles.",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado obtenido correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LibroResponseDTO" }
                },
                "example": [
                  {
                    "id": 1,
                    "titulo": "Don Quijote de la Mancha",
                    "autor": "Miguel de Cervantes",
                    "isbn": "9788424118044",
                    "anioPublicacion": 1605,
                    "disponible": true
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Libros"],
        "summary": "Registrar libro",
        "description": "Crea un libro nuevo. El ISBN debe ser único en el sistema (Regla de Oro).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LibroRequestDTO" },
              "example": {
                "titulo": "Rayuela",
                "autor": "Julio Cortázar",
                "isbn": "9788437604572",
                "anioPublicacion": 1963,
                "disponible": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Libro registrado exitosamente.",
            "headers": {
              "Location": {
                "description": "URI del recurso creado",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LibroResponseDTO" }
              }
            }
          },
          "400": {
            "description": "Datos inválidos en la solicitud.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Conflicto: ya existe un libro con ese ISBN.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/libros/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Identificador numérico del libro",
          "schema": { "type": "integer", "format": "int64" }
        }
      ],
      "get": {
        "tags": ["Libros"],
        "summary": "Obtener libro por ID",
        "responses": {
          "200": {
            "description": "Libro encontrado.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LibroResponseDTO" }
              }
            }
          },
          "404": {
            "description": "Libro no encontrado.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Libros"],
        "summary": "Actualizar libro",
        "description": "Modifica los datos de un libro existente. Aplica las mismas validaciones que el registro y respeta la unicidad de ISBN.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LibroRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Libro actualizado correctamente.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LibroResponseDTO" }
              }
            }
          },
          "400": {
            "description": "Datos inválidos.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Libro no encontrado.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "ISBN duplicado en otro registro.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Libros"],
        "summary": "Eliminar libro",
        "responses": {
          "204": { "description": "Libro eliminado exitosamente (sin contenido)." },
          "404": {
            "description": "Libro no encontrado.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LibroResponseDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "example": 1 },
          "titulo": { "type": "string", "example": "Don Quijote de la Mancha" },
          "autor": { "type": "string", "example": "Miguel de Cervantes" },
          "isbn": { "type": "string", "example": "9788424118044" },
          "anioPublicacion": { "type": "integer", "example": 1605 },
          "disponible": { "type": "boolean", "example": true }
        }
      },
      "LibroRequestDTO": {
        "type": "object",
        "required": ["titulo", "autor", "isbn", "anioPublicacion"],
        "properties": {
          "titulo": {
            "type": "string",
            "minLength": 1,
            "description": "Obligatorio. No puede estar vacío.",
            "example": "Rayuela"
          },
          "autor": {
            "type": "string",
            "minLength": 1,
            "description": "Obligatorio. No puede estar vacío.",
            "example": "Julio Cortázar"
          },
          "isbn": {
            "type": "string",
            "pattern": "^[0-9Xx-]{10,17}$",
            "description": "Entre 10 y 17 caracteres (dígitos, guiones o X). Debe ser único.",
            "example": "9788437604572"
          },
          "anioPublicacion": {
            "type": "integer",
            "minimum": 1001,
            "description": "Debe ser mayor a 1000.",
            "example": 1963
          },
          "disponible": {
            "type": "boolean",
            "description": "Opcional. Si no se envía, el libro se registra como disponible.",
            "example": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "mensaje": { "type": "string", "example": "El libro con el ID 99 no existe en el sistema." },
          "codigoEstado": { "type": "integer", "example": 404 },
          "timestamp": { "type": "integer", "format": "int64", "example": 1747584000000 }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "mensaje": { "type": "string", "example": "La solicitud contiene datos invalidos." },
          "codigoEstado": { "type": "integer", "example": 400 },
          "timestamp": { "type": "integer", "format": "int64" },
          "errores": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "example": {
              "titulo": "El titulo es obligatorio y no puede estar vacio.",
              "anioPublicacion": "El anio de publicacion debe ser mayor a 1000."
            }
          }
        }
      }
    }
  }
}